third_party/blink/web_tests/external/wpt/editing/other/typing-space-in-editable-button.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Type space key in editable element in <button> shouldn't be handled by the <button>
  assert_equals: Click event should not be fired on the <button> expected null but got object "[object PointerEvent]"
Harness: the test ran to completion.
